- Fix: enable advanced (nested) query's
Credits go to: Brian M Hiss see https://groups.google.com/forum/?fromgroups=#!topic/deployd-users/NidQ1sT9u6o
Change:
var prop = props[key]

To:
var prop = props[key.split('.')[0]]

Reason:
If the key is not in the properties, it is excluding it from the query.  We are now checking that at least the first part of the query is defined in properties and allowing it.